Debugging

+++ MCTS_BlokusDuo_cpp/main.cpp
@@ -7,7 +7,7 @@
 #define NUM_THREADS 12
 
 Tree *gameTree;
-int nNodes;
+long nNodes;
 std::mutex mtx;
 
 void expandGameSubtree(Node *node) {
